requirements.txt
# PyDownloader 依赖列表
# 使用清华源安装: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

# UI框架
PySide6>=6.6.0

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# 配置文件
PyYAML>=6.0

# 可选依赖（性能优化）
psutil>=5.9.0

# 开发依赖
pytest>=7.4.0
pytest-qt>=4.2.0
black>=23.0.0
